requirements.txt
ansible==12.0.0
ansible-core==2.19.2
cffi==2.0.0
click==8.3.0
cryptography==46.0.1
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
packaging==25.0
pycparser==2.23
Pygments==2.19.2
PyYAML==6.0.2
resolvelib==1.2.0
rich==14.1.0
shellingham==1.5.4
typer==0.17.4
typing_extensions==4.15.0
